Add tools script to list `torch` APIs
This adds a script that will run through everything in `dir(torch)` recursively and print out each path. We can run this for 1.8.1 and then again for 1.9.0 to determine the API difference between the two in order to ensure everything is properly documented and there are no secret BC-breaking changes.

ghstack-source-id: 6e3e4f948c2ad50a60578ba6c76e44135aed6b80
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58610