Use GitHub's diff directly in clang-tidy
This changes clang-tidy in lint.yml to pull the raw diff from GitHub and parse that rather than use the PRs base revision. The base revision can cause the spurious inclusion of files not changed in the PR as in https://github.com/pytorch/pytorch/pull/59967/checks?check_run_id=2832565901. We could be smarter about how we query git, but this approach ends up being simpler since we just need to search for the diff headers in the .diff file.

ghstack-source-id: 21c92c1129c024bd39dc9c7ec50d72ae4bbeb567
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60048