Sparse CSR: add test parametrization for enabling sparse_csr
This PR adds a new subclass of `_TestParametrizer` that would allow
reusing existing tests by parametrizing whether sparse_csr variant of
the test should be generated or not. This is done by adding a boolean
kwarg `sparse_csr` to the test, which can be used to generate sparse
inputs for the tested functions.
The generated test has name: test_name + 'sparse_csr' + device + dtype.
When `sparse_csr==False` the original test is generated with no
modifications.
`enableSparseCSR` allows specifying CSR specific dtypes and device
skips. Once CSR matrices are better supported the dtype and device
coverage would be the same as for dense tensors.

ghstack-source-id: 90908453795ac8505671cccbd392c271d96f6c43
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61408