[quant][fx] Don't assume bias is a keyword argument
`prepare_fx` currently assumes that bias is always a positional argument to
convolutions, and only a key argument to other functions. This happens to work
today due to a quirk in how `__torch_function__` is handled for python
functions but shouldn't be considered stable.

Instead, we should support `bias` for both positional and keyword forms.

ghstack-source-id: 74c267fb158e2f9a22d9f98c7353e1efa017c158
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61647