[nnc] Insert alloc/free at global scope
Alloc/free inside a loop isn't really an optimization, and furthermore
it breaks some attempted optimization in the llvm backend: we use alloca for
small allocations, which is efficient since alloca is on the stack, but there's
no corresponding free, so we leak tons of stack.  I hit this while building an
rfactor buffer inside a very deeply nested loop.

Differential Revision: [D29533364](https://our.internmc.facebook.com/intern/diff/D29533364/)

ghstack-source-id: 133627310
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61725