ns for fx: return results in execution order
Summary:

By default, NS graph matching matches from the end of the graph
to the start.  This PR reverses the returned results so that
the outputs of the NS APIs are in the order of execution, making
it easier to analyze.

Test Plan:

```
python test/test_quantization.py TestFXGraphMatcher.test_results_order
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 1289a8cd40994dfdeab3421b08cd3405aa5eb05f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61360